+++ PyArr/radarr_api.py
@@ -67,63 +67,25 @@ class RadarrAPI(RequestAPI):
         res = self.request_get(path)
         return res.json()
 
-    def __setCommand(self, data):
-        """Private Command Method
+    def setCommand(self, **kwargs):
+        """Performs any of the predetermined Radarr command routines.
 
-        Args:
-            data (dict): data payload to send to /api/command
+        Kwargs:
+            Required - name (string).
 
+            Options available: RefreshMovie, RescanMovie, MoviesSearch, DownloadedMoviesScan, RssSync, RenameFiles, RenameMovie, CutOffUnmetMoviesSearch, NetImportSync, missingMoviesSearch
+
+            Additional Parameters may be required or optional...
+            See https://github.com/Radarr/Radarr/wiki/API:Command
         Returns:
-            json response
+        json response
+
         """
-        print(data)
         path = "/api/command"
+
+        data = kwargs
         res = self.request_post(path, data)
         return res.json()
-
-    def refreshMovie(self, *args):
-        """RefreshMovie refreshes movie information and rescans disk.
-
-        Args:
-            Optional - movieId (int)
-        Returns:
-            json response
-
-        """
-        data = {}
-        if len(args) == 1:
-            data.update({"name": "RefreshMovie", "movieId": args[0]})
-        else:
-            data.update({"name": "RefreshMovie"})
-        return self.__setCommand(data)
-
-    def rescanMovie(self, *args):
-        """RescanMovie scans disk for any downloaded movie for all or specified movie.
-
-        Args:
-            Optional - movieId (int)
-        Returns:
-            json response
-
-        """
-        data = {}
-        if len(args) == 1:
-            data.update({"name": "RescanMovie", "movieId": args[0]})
-        else:
-            data.update({"name": "RescanMovie"})
-        return self.__setCommand(data)
-
-    def syncRss(self):
-        """Instruct Sonarr to perform an RSS sync with all enabled indexers
-
-        Args:
-            none
-        Returns:
-            json response
-
-        """
-        data = {"name": "RssSync"}
-        return self.__setCommand(data)
 
     def getDiskSpace(self):
         """GetDiskSpace retrieves info about the disk space on the server.
@@ -153,6 +115,19 @@ class RadarrAPI(RequestAPI):
             path = "/api/movie"
 
         res = self.request_get(path)
+        return res.json()
+
+    def updMovie(self, data):
+        """Update an existing movie.
+
+        Args:
+            data (dictionary containing an object obtained by getMovie())
+        Returns:
+            json response
+        """
+
+        path = "/api/movie"
+        res = self.request_put(path, data)
         return res.json()
 
     def lookupMovie(self, term):
@@ -187,12 +162,17 @@ class RadarrAPI(RequestAPI):
         res = self.request_get(path)
         return res.json()
 
-    def constructMovieJson(self, dbId, qualityProfileId):
+    def constructMovieJson(
+        self, dbId, qualityProfileId, rootDir, monitored=True, searchForMovie=True
+    ):
         """Searches for movie on tmdb and returns Movie json to add
 
         Args:
             Required - dbID, <imdb or tmdb id>
             Required - qualityProfileId (int)
+            Required - rootDir (string)
+            Optional - monitored (boolean)
+            Optional - searchForMovie (boolean)
 
         Return:
             JsonArray
@@ -200,32 +180,36 @@ class RadarrAPI(RequestAPI):
         """
         s_dict = self.lookupMovie(dbId)
 
-        root = self.getRoot()[0]["path"]
         movie_json = {
             "title": s_dict["title"],
-            "path": root + s_dict["title"],
+            "path": rootDir + s_dict["title"],
             "qualityProfileId": qualityProfileId,
             "profileId": qualityProfileId,
             "year": s_dict["year"],
             "tmdbId": s_dict["tmdbId"],
             "images": s_dict["images"],
             "titleSlug": s_dict["titleSlug"],
-            "monitored": True,
-            "addOptions": {"searchForMovie": True},
+            "monitored": monitored,
+            "addOptions": {"searchForMovie": searchForMovie},
         }
         return movie_json
 
-    def addMovie(self, dbId, qualityProfileId):
+    def addMovie(
+        self, dbId, qualityProfileId, rootDir, monitored=True, searchForMovie=True
+    ):
         """addMovie adds a new movie to collection
 
         Args:
-            Required - dbid
-            Required - qualityProfileId
+            Required - dbid <imdb or tmdb id>
+            Required - qualityProfileId (int)
+            Required - rootDir (string)
         Returns:
             json response
 
         """
-        movie_json = self.constructMovieJson(dbId, qualityProfileId)
+        movie_json = self.constructMovieJson(
+            dbId, qualityProfileId, rootDir, monitored, searchForMovie
+        )
 
         path = "/api/movie"
         res = self.request_post(path, data=movie_json)